requirements.txt
streamlit>=1.32.0
paddleocr>=2.7.0
paddlepaddle>=2.6.0
transformers>=4.38.0
torch>=2.0.0
torchvision>=0.15.0
rapidfuzz>=3.6.0
opencv-python-headless>=4.9.0
Pillow>=10.2.0
numpy>=1.24.0
pandas>=2.0.0